requirements.txt
numpy==1.26.4

flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

# YouTube ingest
yt-dlp==2024.08.06

# OpenAI SDK
openai==1.40.0

# transcription (local)
faster-whisper==1.0.3

# Fast diarization stack (CPU-friendly, fallback)
webrtcvad==2.0.10
resemblyzer==0.1.4
spectralcluster==0.2.4
pydub==0.25.1

# Fix for OpenAI SDK compatibility
httpx==0.27.2

# Torch CPU build (needed by resemblyzer anyway)
torch==2.2.2

# Optional accurate diarization (pyannote)
pyannote.audio==3.1.1
torchaudio==2.2.2
soundfile==0.12.1
huggingface_hub==0.24.6

reportlab==4.2.2